LPS-120146 Move the configuration file to the proper folder
