+++ modules/apps/data-cleanup/data-cleanup-test/src/testIntegration/java/com/liferay/data/cleanup/test/DataCleanupTest.java
@@ -223,7 +223,8 @@ public class DataCleanupTest {
 	}
 
 	private static final String _CONFIGURATION_PID =
-		"com.liferay.data.cleanup.internal.upgrade.DataCleanupConfiguration";
+		"com.liferay.data.cleanup.internal.configuration." +
+			"DataCleanupConfiguration";
 
 	private static final String[] _SERVLET_CONTEXT_NAMES = {
 		"com.liferay.chat.service", "com.liferay.dictionary.web",
